+++ examples/riverpod_example/lib/router/zones/app_routes.dart
@@ -21,55 +21,39 @@ enum AppRoutes implements DwNavigationRoute<RouterRefreshNotifier> {
   String get zoneRoot => '';
 
   @override
-  List<DwNavigationGuard<RouterRefreshNotifier>> get zoneGuards => [
-    (notifier) => !notifier.isLoggedIn ? AuthRoutes.auth.fullPath : null,
-  ];
+  DwShellRoutePageBuilder? get shellRouteBuilder => null;
 
   @override
-  DwShellRoutePageBuilder? get shellRouteBuilder =>
-      (BuildContext context, GoRouterState state, Widget child) {
+  DwStatefulShellRouteBuilder? get statefulShellRouteBuilder =>
+      (BuildContext context, GoRouterState state, StatefulNavigationShell navigationShell) {
         return DwPageBuilder.fade(
           context,
           state.pageKey,
-          Consumer(
-            child: child,
-            builder: (context, ref, child) {
-              final currentRootRoute = ref
-                  .read(appRouterProvider)
-                  .rootRouteFromState(state);
-
-              final currentIndex = currentRootRoute == AppRoutes.profile
-                  ? 1
-                  : 0;
-
-              return Scaffold(
-                body: child,
-                bottomNavigationBar: BottomNavigationBar(
-                  currentIndex: currentIndex,
-                  onTap: (index) {
-                    switch (index) {
-                      case 0:
-                        context.goNamed(AppRoutes.catalog.name);
-                        break;
-                      case 1:
-                        context.goNamed(AppRoutes.profile.name);
-                        break;
-                    }
-                  },
-                  items: const [
-                    BottomNavigationBarItem(
-                      icon: Icon(Icons.list),
-                      label: 'Catalog',
-                    ),
-                    BottomNavigationBarItem(
-                      icon: Icon(Icons.person),
-                      label: 'Profile',
-                    ),
-                  ],
+          Scaffold(
+            body: navigationShell,
+            bottomNavigationBar: BottomNavigationBar(
+              currentIndex: navigationShell.currentIndex,
+              onTap: (i) => navigationShell.goBranch(
+                i,
+                initialLocation: i == navigationShell.currentIndex,
+              ),
+              items: const [
+                BottomNavigationBarItem(
+                  icon: Icon(Icons.list),
+                  label: 'Catalog',
                 ),
-              );
-            },
+                BottomNavigationBarItem(
+                  icon: Icon(Icons.person),
+                  label: 'Profile',
+                ),
+              ],
+            ),
           ),
         );
       };
+
+  @override
+  List<DwNavigationGuard<RouterRefreshNotifier>> get zoneGuards => [
+    (notifier) => !notifier.isLoggedIn ? AuthRoutes.auth.fullPath : null,
+  ];
 }